solver: remove old solver interface and rename new one
