+++ conjure_oxide/src/lib.rs
@@ -6,7 +6,6 @@ pub mod generate_custom;
 pub mod parse;
 pub mod rule_engine;
 pub mod rules;
-pub mod solvers;
 pub mod utils;
 
 pub use conjure_core::ast; // re-export core::ast as conjure_oxide::ast
@@ -16,4 +15,7 @@ pub use rules::eval_constant;
 
 pub use error::Error;
 
+pub mod solver;
+
+#[doc(hidden)]
 pub mod unstable;
