+++ conjure_oxide/src/solver/adaptors/mod.rs
