+++ conjure_oxide/src/solver/stats.rs
