--- conjure_oxide/src/solvers/mod.rs
@@ -1,13 +1,0 @@
-mod error;
-pub mod kissat;
-pub mod minion;
-
-pub use crate::ast::Model;
-pub use error::*;
-
-pub trait FromConjureModel
-where
-    Self: Sized,
-{
-    fn from_conjure(model: Model) -> Result<Self, SolverError>;
-}
