+++ conjure_oxide/src/unstable/mod.rs
@@ -1,3 +1,1 @@
 //! Unstable and in-development features of Conjure-Oxide.
-
-pub mod solver_interface;
